+++ flask_restplus/reqparse.py
@@ -29,6 +29,7 @@ class ParseResult(dict):
     def __setattr__(self, name, value):
         self[name] = value
 
+
 _friendly_location = {
     'json': 'the JSON body',
     'form': 'the post body',
@@ -177,7 +178,7 @@ class Argument(object):
         errors = {self.name: error_msg}
 
         if bundle_errors:
-            return error, errors
+            return ValueError(error), errors
         abort(400, 'Input payload validation failed', errors=errors)
 
     def parse(self, request, bundle_errors=False):
@@ -229,9 +230,7 @@ class Argument(object):
 
                     if self.choices and value not in self.choices:
                         msg = '{0} is not a valid choice'.format(value)
-                        if bundle_errors:
-                            return self.handle_validation_error(msg, bundle_errors)
-                        self.handle_validation_error(msg, bundle_errors)
+                        return self.handle_validation_error(msg, bundle_errors)
 
                     if name in request.unparsed_arguments:
                         request.unparsed_arguments.pop(name)
@@ -244,9 +243,7 @@ class Argument(object):
                 locations = [_friendly_location.get(loc, loc) for loc in self.location]
                 location = ' or '.join(locations)
             error_msg = 'Missing required parameter in {0}'.format(location)
-            if bundle_errors:
-                return self.handle_validation_error(error_msg, bundle_errors)
-            self.handle_validation_error(error_msg, bundle_errors)
+            return self.handle_validation_error(error_msg, bundle_errors)
 
         if not results:
             if callable(self.default):
@@ -363,7 +360,7 @@ class RequestParser(object):
             if found or arg.store_missing:
                 result[arg.dest or arg.name] = value
         if errors:
-            abort(400, message=errors)
+            abort(400, 'Input payload validation failed', errors=errors)
 
         if strict and req.unparsed_arguments:
             arguments = ', '.join(req.unparsed_arguments.keys())